+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
  - `lewd.integer` no longer coerces strings.
  
+###### New features:
+
+ - Check for distinct values in an array with `lewd.unique`.
+ 
 ###### Improvements:
 
  - Support for an even more concise notation for `lewd.integer` and `lewd.isoDateTime`. 
-
+ - Better error handling when using conditions with incorrect parameters.
 
 ## 0.3.0 (2014-06-03)